Run linter + fix hydration issue + add filter btn
Hydration issue is only half fixed. The only thing that works is to make
it client-only, but that means it won't be seen by search engines (SEO).
This isn't a big problem as these are only the thumbnails of the
projects containing not to much information. However it has to be kept
in mind for later and see if it has an effect on SEO.